Unify REST/WS fill protocol type, clarify quantity signedness (#1037)
Closes A-2539

> Our sign and size conventions for quantities, etc. are a mess. We
ought to conform our protocol types to what is canonical/colloquial in
trading APIs, aka preference for (unsigned quantity, side/direction)
over signed quantity. This also harmonizes quantity fields at 64-bit
wide integers.

## Changelog

- BREAKING CHANGE: Protocol `Fill.quantity` is now unsigned (`u64`) with
a new required `side` field (`Side`)
- BREAKING CHANGE: `Position` fields renamed from
`open_quantity`/`open_notional` to `signed_quantity`/`signed_notional`
- BREAKING CHANGE: `SymbolRiskSnapshot` fields renamed from
`open_quantity`/`open_notional` to `signed_quantity`/`signed_notional`
- BREAKING CHANGE: Order quantities changed from `i32` to `u64`
(`PlaceOrderRequest.quantity`,
`OrderDetails.quantity/filled_quantity/remaining_quantity`,
`OrderStatus.filled_quantity/remaining_quantity`)
- BREAKING CHANGE: Ticker quantities changed to `u64`
(`last_trade_quantity`, `total_volume`)
- BREAKING CHANGE: Book level quantities changed to `u64`
(`L2BookLevel.quantity`, `L3BookLevel.quantity/order_quantities`,
`GetBookResponseBookLevel.total_quantity/orders`)
- BREAKING CHANGE: Candle volumes changed to `u64` (`buy_volume`,
`sell_volume`, `volume`)
- BREAKING CHANGE: `PreviewAggressiveLimitOrderRequest.quantity` and
response quantities changed to `u64`

## Plan

### Overview

This plan establishes a consistent convention across the codebase:
- All `quantity` and `volume` fields should be **unsigned 64-bit**
(`u64`)
- Signed quantities should be explicitly named `signed_quantity` with
type `i64`

### Guiding Principles

1. **Quantities are magnitudes**: A quantity represents "how much" and
is inherently non-negative
2. **Direction is separate**: Side (Buy/Sell) or PositionSide
(Long/Short) is orthogonal to quantity
3. **Explicit over implicit**: If a field encodes direction in its sign,
it should be named `signed_quantity`
4. **64-bit throughout**: All quantities and volumes use 64-bit integers
for consistency with EP3

#### Why 64-bit?

- EP3 (the exchange protocol) uses 64-bit integers - matching this
avoids range/conversion issues
- Consistent width throughout the codebase reduces confusion
- No need to distinguish between "quantity" (32-bit) and "volume"
(64-bit)

---------

+++ src/protocol/marketdata_publisher.rs
@@ -72,7 +72,7 @@ pub struct Ticker {
     pub last_trade_price: Option<Decimal>,
     /// Last trade quantity in contracts
     #[serde(rename = "q")]
-    pub last_trade_quantity: i32,
+    pub last_trade_quantity: u64,
     /// Session open price in USD
     #[serde(rename = "o")]
     pub session_open_price: Option<Decimal>,
@@ -84,10 +84,10 @@ pub struct Ticker {
     pub session_high_price: Option<Decimal>,
     /// Total volume in contracts
     #[serde(rename = "v")]
-    pub total_volume: i32,
+    pub total_volume: u64,
     /// Open interest in contracts
     #[serde(rename = "oi")]
-    pub open_interest: u32,
+    pub open_interest: u64,
     /// Instrument state
     #[serde(rename = "i")]
     #[serde(default)]
@@ -131,7 +131,7 @@ pub struct L2BookLevel {
     #[serde(rename = "p")]
     pub price: Decimal,
     #[serde(rename = "q")]
-    pub quantity: i32,
+    pub quantity: u64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -140,9 +140,9 @@ pub struct L3BookLevel {
     #[serde(rename = "p")]
     pub price: Decimal,
     #[serde(rename = "q")]
-    pub quantity: i32,
+    pub quantity: u64,
     #[serde(rename = "o")]
-    pub order_quantities: Vec<i32>,
+    pub order_quantities: Vec<u64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -151,7 +151,7 @@ pub struct Trade {
     #[serde(rename = "p")]
     pub price: Decimal,
     #[serde(rename = "q")]
-    pub quantity: i32,
+    pub quantity: u64,
     #[serde(rename = "s")]
     pub symbol: String,
     #[serde(rename = "d")]
